editor/composer/test/test_bug738440.html

<!doctype html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=738440
-->
<title>Test for Bug 738440</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css" />
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=738440">Mozilla Bug 738440</a>
<div contenteditable></div>
<script>

/** Test for Bug 738440 **/
document.execCommand("stylewithcss", false, "true");
is(document.queryCommandState("stylewithcss"), true,
   "setting stylewithcss to true should cause its state to be true");
is(document.queryCommandState("usecss"), false,
   "usecss state should always be false");

document.execCommand("stylewithcss", false, "false");
is(document.queryCommandState("stylewithcss"), false,
   "setting stylewithcss to false should cause its state to be false");
is(document.queryCommandState("usecss"), false,
   "usecss state should always be false");

document.execCommand("usecss", false, "true");
is(document.queryCommandState("stylewithcss"), false,
   "setting usecss to true should cause stylewithcss state to be false");
is(document.queryCommandState("usecss"), false,
   "usecss state should always be false");

document.execCommand("usecss", false, "false");
is(document.queryCommandState("stylewithcss"), true,
   "setting usecss to false should cause stylewithcss state to be true");
is(document.queryCommandState("usecss"), false,
   "usecss state should always be false");

</script>
